backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
pyjwt==2.8.0
cryptography==42.0.7
python-multipart==0.0.9
